вынес функцию get_or_create в models
